Remove household members array data structure
+ Not being used in app, so let's delete -- keep complexity down
+ Move all logic to the DocumentsApi class, up one level in the heirarchy
+ Also remove IdentityDocuments class because it became stubby
