+++ documents_api.rb
@@ -1,4 +1,4 @@
-require_relative "models/household_member"
+require_relative "load_documents_data"
 require_relative "models/residency_documents"
 require_relative "helpers/string_parser"
 require          "active_support/all"
@@ -6,24 +6,34 @@ require          "active_support/all"
 module Api
   class DocumentsRequest
 
-    def initialize(household_members:,
-                   has_rental_income:,
+    def initialize(has_rental_income:,
                    renting:,
                    owns_home:,
                    shelter:,
                    living_with_family_or_friends:,
-                   all_citizens:
-    )
+                   all_citizens:,
+                   employee:,
+                   disability_benefits:,
+                   child_support:,
+                   self_employed:,
+                   retired:,
+                   unemployment_benefits:
+      )
 
-      @household_member = get_household_member_data(household_members)
       @has_rental_income = StringParser.new(has_rental_income).to_boolean
       @renting = StringParser.new(renting).to_boolean
       @owns_home = StringParser.new(owns_home).to_boolean
       @shelter = StringParser.new(shelter).to_boolean
       @living_with_family_or_friends = StringParser.new(living_with_family_or_friends).to_boolean
       @all_citizens = StringParser.new(all_citizens).to_boolean
+      @employee = StringParser.new(employee).to_boolean
+      @self_employed = StringParser.new(self_employed).to_boolean
+      @disability_benefits = StringParser.new(disability_benefits).to_boolean
+      @child_support = StringParser.new(child_support).to_boolean
+      @retired = StringParser.new(retired).to_boolean
+      @unemployment_benefits = StringParser.new(unemployment_benefits).to_boolean
 
-      raise "Badly formatted request" if @has_rental_income.nil?
+      raise "Invalid data" unless valid_data?
     end
 
     def fetch_documents
@@ -35,6 +45,36 @@ module Api
       }
     end
 
+    private
+
+    ## DATA VALIDATION ##
+
+    def valid_data?
+      attribute_classes.all? do |class_type|
+        (class_type == TrueClass) || (class_type == FalseClass)
+      end
+    end
+
+    def attribute_classes
+      [
+        @has_rental_income,
+        @renting,
+        @owns_home,
+        @shelter,
+        @living_with_family_or_friends,
+        @all_citizens,
+        @employee,
+        @self_employed,
+        @disability_benefits,
+        @child_support,
+        @retired,
+        @unemployment_benefits,
+      ].map { |attribute| attribute.class }
+    end
+
+
+    ## RESIDENCY DOCUMENTS ##
+
     def residency_documents
       return ResidencyDocuments.new(
         renting: @renting,
@@ -44,26 +84,74 @@ module Api
       ).documents
     end
 
-    def identity_documents
-      @household_member.documents_based_on_identity
+
+    ## IDENTITY DOCUMENTS ##
+
+    def needs_identity_docs
+      !@employee &&
+      !@self_employed &&
+      !@disability_benefits &&
+      !@child_support &&
+      !@unemployment_benefits
     end
+
+    def identity_documents
+      needs_identity_docs ? identity_documents_list : []
+    end
+
+    def identity_documents_list
+      [
+        STATE_ID,
+        SCHOOL_PHOTO_ID,
+        US_MILITARY_CARD,
+        VOTER_REGISTRATION_CARD,
+        BIRTH_CERTIFICATE
+      ]
+    end
+
+
+    ## CITIZENSHIP DOCUMENTS ##
 
     def citizenship_documents
       @all_citizens ? [] : [I_90_DOCUMENTATION]
     end
 
+
+    ## INCOME DOCUMENTS ##
+
     def income_documents
-      @household_member.documents_based_on_income
+      [
+        documents_based_on_employment,
+        documents_based_on_self_employment,
+        documents_based_on_child_support,
+        documents_based_on_disability,
+        documents_based_on_retirement,
+        documents_based_on_unemployment,
+      ].compact
     end
 
-    def get_household_member_data(data)
-      household_member_hash = data.to_a[0][1].symbolize_keys!
+    def documents_based_on_employment
+      PAY_STUBS if @employee
+    end
 
-      household_member_hash.each_key do |k|
-        household_member_hash[k] = StringParser.new(household_member_hash[k]).to_boolean
-      end
+    def documents_based_on_self_employment
+      SELF_EMPLOYMENT_FORM if @self_employed
+    end
 
-      return HouseholdMember.new(household_member_hash)
+    def documents_based_on_child_support
+      WRITTEN_CHILD_SUPPORT_STATEMENT if @child_support
+    end
+
+    def documents_based_on_disability
+      AWARD_LETTER_FOR_DISABILITY if @disability_benefits
+    end
+
+    def documents_based_on_retirement
+      AWARD_LETTER_FOR_SOCIAL_SECURITY if @retired
+    end
+
+    def documents_based_on_unemployment
+      AWARD_LETTER_FOR_UNEMPLOYMENT if @unemployment_benefits
     end
 
   end
