--- models/identity_documents.rb
@@ -1,13 +1,0 @@
-class IdentityDocuments
-
-  def self.documents
-    return [
-      STATE_ID,
-      SCHOOL_PHOTO_ID,
-      US_MILITARY_CARD,
-      VOTER_REGISTRATION_CARD,
-      BIRTH_CERTIFICATE
-    ]
-  end
-
-end
