+++ public/javascripts/components/default_data.js
@@ -2,16 +2,12 @@
   window.shared || (window.shared = {});
 
   window.shared.DefaultData = {
-    "household_members": [
-      {
-        "disability_benefits": "false",
-        "employee": "false",
-        "self_employed": "false",
-        "child_support": "false",
-        "retired": "false",
-        "unemployment_benefits": "false",
-      }
-    ],
+    "disability_benefits": "false",
+    "employee": "false",
+    "self_employed": "false",
+    "child_support": "false",
+    "retired": "false",
+    "unemployment_benefits": "false",
     "has_rental_income": "false",
     "renting": "false",
     "owns_home": "false",
